% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KafkaProxyClient.R
\name{GetOrCreateConsumerInstance}
\alias{GetOrCreateConsumerInstance}
\title{Get Or Create Pharbers/Blackmirror Comsumer Instance}
\usage{
GetOrCreateConsumerInstance()
}
\description{
Get Or Create Pharbers/Blackmirror Comsumer Instance
}
