% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KafkaProxyClient.R
\name{PushMessage}
\alias{PushMessage}
\title{Push R message to pharber/blackmirror message queue}
\usage{
PushMessage(content)
}
\arguments{
\item{content}{the message json buffer pushed to the kafka}
}
\description{
Push R message to pharber/blackmirror message queue
}
